delete unused service example
