--- internal/service/my_service/my_service.go
@@ -1,8 +1,0 @@
-package my_service
-
-type MyService struct {
-}
-
-func (m *MyService) Find() {
-
-}
